Organizacao pastas e inicio do PUT e POST das TASKS
