+++ src/pages/Engenheiro/Projeto/index.jsx
@@ -3,8 +3,8 @@ import { useParams } from "react-router-dom";
 import BodyHeader from "@/components/BodyHeader";
 import TabelaWbs from "@/components/TabelaWbs";
 import TabelaCronograma from "@/components/TabelaCronograma";
-import Dashboard from "../../components/Dashboard/Dashboard";
-import LiderSelect from "../../components/LiderSelect/LiderSelect";
+import Dashboard from "../../../components/Dashboard/Dashboard";
+import LiderSelect from "../../../components/LiderSelect/LiderSelect";
 
 const Projeto = () => {
   const { id, itemId } = useParams();
